third_party/blink/web_tests/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-vertical-writing-mode-navigation.optional-expected.txt

This is a testharness.js-based test.
[FAIL] vertical-lr: Right/Left keys navigate as primary axis
  assert_equals: ArrowRight should move to the next option. expected Element node <option class="two">two</option> but got Element node <option class="one">one</option>
[FAIL] vertical-rl: Left/Right keys navigate as primary axis (block-flipped)
  assert_equals: ArrowLeft should move to the next option. expected Element node <option class="two">two</option> but got Element node <option class="one">one</option>
[FAIL] sideways-lr: Right/Left keys navigate as primary axis
  assert_equals: ArrowRight should move to the next option. expected Element node <option class="two">two</option> but got Element node <option class="one">one</option>
[FAIL] sideways-rl: Left/Right keys navigate as primary axis (block-flipped)
  assert_equals: ArrowLeft should move to the next option. expected Element node <option class="two">two</option> but got Element node <option class="one">one</option>
Harness: the test ran to completion.
